requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.7.4.post0=pypi_0
aiozipstream=0.4=pypi_0
alembic=1.4.2=pypi_0
appdirs=1.4.4=py_0
appnope=0.1.0=pypi_0
apscheduler=3.7.0=pypi_0
arrow=0.15.6=pypi_0
astroid=2.4.1=py37_0
async-timeout=3.0.1=pypi_0
asyncinit=0.2.4=pyhd8ed1ab_0
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.5.7=pypi_0
bcrypt=3.1.7=pypi_0
beautifulsoup4=4.9.3=pypi_0
cchardet=2.1.7=pypi_0
certifi=2021.5.30=py37hecd8cb5_0
cffi=1.14.5=py37h2125817_0
chardet=4.0.0=py37hecd8cb5_1003
cssselect=1.1.0=pypi_0
cssutils=1.0.2=pypi_0
dependency-injector=4.34.0=pypi_0
devtools=0.6.1=pypi_0
docutils=0.15.2=pypi_0
faker=4.1.0=pypi_0
fastapi=0.55.1=pypi_0
fastapi-sqlalchemy=0.1.4=pypi_0
gne=0.2.6=pypi_0
h11=0.9.0=pypi_0
httptools=0.1.1=pypi_0
importlib-metadata=4.6.0=pypi_0
libcxx=10.0.0=1
libedit=3.1.20181209=hb402a30_0
libffi=3.3=h0a44026_1
libiconv=1.16=h1de35cc_0
libxml2=2.9.10=h7cdb67c_3
libxslt=1.1.34=h83b36ba_0
lxml=4.5.1=pypi_0
motor=2.4.0=pypi_0
packaging=20.4=pypi_0
passlib=1.7.2=pypi_0
pickleshare=0.7.5=pypi_0
pip=20.0.2=py37_3
ptyprocess=0.6.0=pypi_0
pydantic=1.5.1=pypi_0
pydantic-yaml=0.4.0=pypi_0
pymongo=3.11.4=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pyppeteer=0.2.2=py_1
pyquery=1.4.3=py37hecd8cb5_0
pytest=5.4.2=pypi_0
pytest-asyncio=0.15.1=pypi_0
python=3.7.7=hf48f09d_4
pytz=2020.1=pypi_0
pyyaml=5.4.1=pypi_0
requests=2.25.1=pyhd3eb1b0_0
setuptools=46.4.0=py37_0
singleton-decorator=1.0.0=pypi_0
six=1.15.0=pypi_0
starlette=0.13.2=pypi_0
typing=3.7.4.1=pypi_0
typing-extensions=3.10.0.0=pypi_0
tzlocal=2.1=pypi_0
urllib3=1.25.9=pypi_0
us=2.0.2=pypi_0
uvicorn=0.11.5=pypi_0
uvloop=0.14.0=pypi_0
wcwidth=0.2.5=pypi_0
websockets=8.1=py37hf967b71_3
wheel=0.34.2=py37_0
wrapt=1.11.2=py37h1de35cc_0
xz=5.2.5=h1de35cc_0
yarl=1.6.3=pypi_0
